docs: Renamed directory features/ to redux-features/ for clarity
docs: Added SettingsFab.scss and SettingsFab.tsx to depreciated/
- Removed extraneous and unclear spacing from all necessary .scss files
