+++ src/components/.depreciated/SettingsFab.scss
@@ -1,4 +1,4 @@
-.FabButton{
+.FabButton {
     --box-shadow: #2e3436;
     --color: var(--ion-color-dark);
     --background: none;
@@ -6,6 +6,4 @@
     --background-hover-opacity: 10%;
     --background-activated-color: none;
     --ripple-color: #4d5d84;
-    
-    
 }
