+++ src/components/FooterBar.scss
@@ -1,23 +1,18 @@
 .FooterToolbar {
     min-width: 100%;
-
     min-height: 15.5%;
     max-height: 31%;
-
     background-color: #000;
 }
 
 .FooterBarRow {
     min-width: 78%;
-
     vertical-align: middle;
-
     row-gap: 11.5%;
 }
 
 .FooterBarGrid {
     text-align: center;
-
     padding-left: 11%;
     padding-right: 11%;
 }
