+++ src/pages/ComponentModal.tsx
@@ -3,7 +3,7 @@ import { IonContent, IonModal } from '@ionic/react';
 import { useDispatch, useSelector } from 'react-redux';
 import { Pages } from '../Models/Enums';
 import { selectRouter } from '../app/reducersindex';
-import { setPage } from '../features/Routing/RouterStore';
+import { setPage } from '../redux-features/Routing/RouterStore';
 import Music from './Music';
 import Car from './Car';
 import Climate from './Climate';
